%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{GenePixResultFile}
\alias{GenePixResultSet}
\alias{as.character.GenePixResultFile}
\alias{byPath.GenePixResultSet}
\alias{getChipType.GenePixResultFile}
\alias{getChipType.GenePixResultSet}
\alias{getExtensionPattern.GenePixResultFile}
\alias{getPlatform.GenePixResultFile}
\alias{getPlatform.GenePixResultSet}
\alias{verify.GenePixResultFile}
\alias{byName.GenePixResultSet}
\alias{findByName.GenePixResultSet}
\alias{readLimmaRGList}
\alias{readLimmaRGList.GenePixResultSet}
\alias{asAMArray}
\alias{asAMArray.RGArray}
\alias{extractAMArray}
\alias{extractAMArray.GenePixResultSet}
\alias{extractRGArray}
\alias{extractRGArray.GenePixResultSet}
\alias{extractRGArray.RGList}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
